+++ tests/page/page-aria-snapshot.spec.ts
@@ -15,20 +15,7 @@
  */
 
 import type { Locator } from '@playwright/test';
-import { test as it, expect } from './pageTest';
-
-export function unshift(snapshot: string): string {
-  const lines = snapshot.split('\n');
-  let whitespacePrefixLength = 100;
-  for (const line of lines) {
-    if (!line.trim())
-      continue;
-    const match = line.match(/^(\s*)/);
-    if (match && match[1].length < whitespacePrefixLength)
-      whitespacePrefixLength = match[1].length;
-  }
-  return lines.filter(t => t.trim()).map(line => line.substring(whitespacePrefixLength)).join('\n');
-}
+import { test as it, expect, unshift } from './pageTest';
 
 async function checkAndMatchSnapshot(locator: Locator, snapshot: string) {
   expect.soft(await locator.ariaSnapshot()).toBe(unshift(snapshot));
