+++ tests/page/pageTest.ts
@@ -54,6 +54,19 @@ export function roundBox(box: BoundingBox): BoundingBox {
   };
 }
 
+export function unshift(snapshot: string): string {
+  const lines = snapshot.split('\n');
+  let whitespacePrefixLength = 100;
+  for (const line of lines) {
+    if (!line.trim())
+      continue;
+    const match = line.match(/^(\s*)/);
+    if (match && match[1].length < whitespacePrefixLength)
+      whitespacePrefixLength = match[1].length;
+  }
+  return lines.filter(t => t.trim()).map(line => line.substring(whitespacePrefixLength)).join('\n');
+}
+
 export const expect = baseExpect.extend({
   toContainYaml(received: string, expected: string) {
     const trimmed = expected.split('\n').filter(a => !!a.trim());
